Use namespace for class
