+++ .scrutinizer.yml
@@ -1,6 +1,6 @@
 filter:
     paths:
-        - 'class-mundschenk-wp-requirements.php'
+        - 'class-wp-requirements.php'
         - 'partials/'
     dependency_paths:
         - wordpress/
@@ -27,11 +27,11 @@ build:
                 override: true
             dependencies:
                 before:
-                    - composer require --dev johnpbloch/wordpress
+                    - composer require --dev johnpbloch/wordpress-core
             tests:
                 override:
                     - php-scrutinizer-run --disable-security-analysis
-                    - phpcs-run --standard=phpcs.xml class-mundschenk-wp-requirements.php partials/ --extensions=php
+                    - phpcs-run --standard=phpcs.xml class-wp-requirements.php partials/ --extensions=php
 
 checks:
     php:
