+++ composer.json
@@ -29,7 +29,7 @@
     "prefer-stable": true,
 
     "autoload": {
-        "classmap": ["class-mundschenk-wp-requirements.php"]
+        "classmap": ["class-wp-requirements.php"]
     },
     "autoload-dev": {
         "classmap": ["tests/"]
@@ -43,7 +43,7 @@
             "@composer test -- --coverage-html tests/coverage"
         ],
         "phpcs": [
-            "vendor/bin/phpcs -p -s class-mundschenk-wp-requirements.php tests/ --extensions=php"
+            "vendor/bin/phpcs -p -s class-wp-requirements.php tests/ --extensions=php"
         ]
     }
 }
